ConcurrencyLevel and ThreadCount PG properties ara converted to uint

+++ src/Inceptum.Messaging/Configuration/ProcessingGroupConfigurationElement.cs
@@ -11,7 +11,7 @@ namespace Inceptum.Messaging.Configuration
             set { this["concurrencyLevel"] = value; }
         }
 
-        [ConfigurationProperty("queueCapacity", IsRequired = false, IsKey = false, DefaultValue = 1024)]
+        [ConfigurationProperty("queueCapacity", IsRequired = false, IsKey = false, DefaultValue = (uint)1024)]
         public uint QueueCapacity
         {
             get { return (uint)this["queueCapacity"]; }
